chore:Collected Interfeces and Modified FC
