+++ frontend/components/layout.tsx
@@ -3,11 +3,11 @@ import Image from "next/image";
 import styles from "./layout.module.css";
 import utilStyles from "../styles/utils.module.css";
 import Link from "next/link";
-
+import { LayoutProps } from "../interface/interface";
 const name = "Choose the Chat Mode";
 export const siteTitle = "Jinora";
 
-export default function Layout({ children, home }) {
+export default function Layout({ children, home }:LayoutProps) {
   return (
     <div className={styles.container}>
       <Head>
