+++ frontend/interface/interface.ts
@@ -1,0 +1,105 @@
+export interface SettingsPopupProps {
+    onClose: () => void;           
+    soundEnabled: boolean;           
+    setSoundEnabled: (enabled: boolean) => void; 
+    notificationsEnabled: boolean;  
+    setNotificationsEnabled: (enabled: boolean) => void; 
+  }
+  
+export interface RightPaneProps extends SettingsPopupProps {} 
+  
+  interface IconProps {
+    src: string;
+    alt: string;
+    onClick?: () => void; 
+  }
+  
+export interface LayoutProps {
+    children: React.ReactNode;
+    home: boolean;
+}
+
+export interface TopicDropdownProps {
+    topic: string;
+    setTopic: (topic: string) => void;
+    login: boolean;
+  }
+  
+export interface Project {
+    Category: string;
+    Name: string;
+    ShortDesc: string;
+    LongDesc: string;
+    ImageLink: string;
+    AppStoreLink: string;
+    GithubLink: string;
+    PlayStoreLink: string;
+  }
+
+  export interface TopicSelectionModalProps {
+    onClose: () => void;
+  }
+
+  export interface UsernameInputProps {
+    value: string;
+    onChange: (value: string) => void;
+  }
+
+export interface NavbarProps{
+    currentPage: string;
+    currentTopic:string;
+  }
+
+export interface ModalProps {
+    isOpen: boolean;
+    onClose: () => void;
+    children: React.ReactNode;
+  }
+
+  export interface BoxProps {
+    channel: "public" | "private" | "chatbot"; 
+  }
+
+export interface MailProps {
+    isOpen?: boolean;
+    onClose: () => void;
+    channel: "public" | "private" | "chatbot";
+  }
+  export interface ProjectListProps {
+    projects: Project[];
+    category: string;
+    heightDecrease?: boolean;
+  }
+
+  export interface ProjectCardProps {
+    name: string;
+    shortDesc: string;
+    ImageLink: string;
+    Github?: string;
+    PlayStore?: string;
+    AppStore?: string;
+  }
+
+  export interface ChatBotLoginModalProps {
+    onClose: () => void; // Function to close the modal
+  }
+  export type Topic = "SELECT A TOPIC" | "Option 1" | "Option 2" | string ;
+
+  export interface ChatContainerProps {
+    messages: string[]; // Assuming messages are serialized JSON strings
+    messagesEndRef: React.RefObject<HTMLDivElement>; // Ref to scroll to the bottom
+  }
+  
+  export interface Message {
+    isSent: boolean;
+    username: string;
+    text: string; // Can contain JSON data in some cases
+  }
+  
+  export interface Message {
+    isSent: boolean;
+    avatar?: string; 
+    username: string;
+    text: string;
+    timestamp: number;
+  }
